Overhaul of the skcosmo folder structure
Main changes:
 - `pcovr` and `kpcovr` are now housed in the `decomposition` folder
 - `orthogonalizers` and `pcovr_utils` (pcovr_covariance and pcovr_kernel) are now housed in the `utils` folder
 - there are now dedicated `feature_selection` and `sample_selection` folders. Given that re-writing the selection modules is underway, these two modules currently point to the `selection` folder. This `selection` folder will be deprecated once the new selectors are written.
 - all links and documentation have been updated
