+++ docs/source/index.rst
@@ -31,6 +31,6 @@ it you agree to distribute your changes under the same license.
   pcovr
   gfrm
   datasets
-  overlap
+  utils
   reference
   contributing
