--- docs/source/overlap.rst
@@ -1,4 +1,0 @@
-Random Partitioning with Overlaps
-=================================
-
-.. automodule:: skcosmo.model_selection._split
