+++ skcosmo/utils/__init__.py
@@ -4,5 +4,13 @@ used by multiple packages
 """
 
 from .progress_bar import get_progress_bar
+from .pcovr_utils import pcovr_covariance, pcovr_kernel
+from .orthogonalizers import feature_orthogonalizer, sample_orthogonalizer
 
-__all__ = ["get_progress_bar"]
+__all__ = [
+    "get_progress_bar",
+    "pcovr_covariance",
+    "pcovr_kernel",
+    "feature_orthogonalizer",
+    "sample_orthogonalizer",
+]
