+++ tests/test_pcovr.py
@@ -1,5 +1,5 @@
 import unittest
-from skcosmo.pcovr import PCovR
+from skcosmo.decomposition import PCovR
 from sklearn.datasets import load_boston
 import numpy as np
 from sklearn import exceptions
